Autosahayak 2.0/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
pydantic==2.11.3
email-validator==2.2.0
jinja2==3.1.6
python-multipart==0.0.20
faiss-cpu==1.10.0
openai==1.76.0
